agent/requirements.txt
requests==2.31.0
asyncio==3.4.3
aiohttp==3.9.1
paramiko==3.4.0
pydantic==2.9.0
python-dotenv==1.0.0
